requirement.txt
androguard==3.3.5
numpy==1.19.3
pyyaml==5.3.1
requests==2.25.0